Refactor code for anime section
